refactor: doc location
