+++ src/app/activities/[id]/_components/ActivityHeader.tsx
@@ -1,6 +1,6 @@
 import { MapPin, Star, EllipsisVertical } from "lucide-react";
 import { Button } from "@/components/ui/Buttons/Button";
-import type { Activity } from "@/commons/consts/activity";
+import type { Activity } from "@/types/activities";
 
 interface ActivityHeaderProps {
   activity: Activity;
