+++ src/types/activities.ts
@@ -1,0 +1,50 @@
+export interface SubImage {
+  id: number;
+  imageUrl: string;
+}
+
+export interface Schedule {
+  id: number;
+  date: string;
+  startTime: string;
+  endTime: string;
+}
+
+export interface Activity {
+  id: number;
+  userId: number;
+  title: string;
+  description: string;
+  category: string;
+  price: number;
+  address: string;
+  bannerImageUrl: string;
+  subImages: SubImage[];
+  schedules: Schedule[];
+  reviewCount: number;
+  rating: number;
+  createdAt: string;
+  updatedAt: string;
+}
+
+export interface ReviewUser {
+  profileImageUrl: string;
+  nickname: string;
+  id: number;
+}
+
+export interface Review {
+  id: number;
+  user: ReviewUser;
+  activityId: number;
+  rating: number;
+  content: string;
+  createdAt: string;
+  updatedAt: string;
+}
+
+export interface Reviews {
+  averageRating: number;
+  totalCount: number;
+  reviews: Review[];
+}
